refactor: extract useTaskDerived hook
